deleted hypergrid.py and changed type of errors
